requirements.txt
body-visualizer @ git+https://github.com/nghorbani/body_visualizer@371d0936479e12b485161c32732c0fe97574e3a0
smplx @ git+https://github.com/vchoutas/smplx@85f8570a4011489985e647a469fd6cc8e96cc68b
boto==2.49.0
colour==0.1.5
chumpy==0.70
flatten-dict==0.4.2
h5py==2.10.0
ipdb==0.13.9
loguru==0.6.0
matplotlib==3.5.0
networkx==2.2
numpy==1.21.5
omegaconf==2.1.0
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
pyopengl
pandas==1.3.4
pathlib
Pillow==8.4.0
pyrender==0.1.33
pytorch3d==0.3.0
scikit-image==0.18.3
scipy==1.2.3
shortuuid==1.0.8
spacepy
tensorboard
torchgeometry==0.1.2
torchmetrics==0.4.1
tqdm
trimesh==3.10.1
